+++ src/Event/Fcl/fclShop.c
@@ -2585,7 +2585,7 @@ u8 * FUN_003e8860(u32 param_1,u32 param_2)
 
   u32 uVar1;
 
-  u8 bVar2;
+  int bVar2;
 
   int *piVar3;
